:art: Add icon for Plutograph and improve file structure
