+++ android/app/src/main/java/com/fra/plutomierz/ui/components/MotivationalText.kt
@@ -1,4 +1,4 @@
-package com.fra.plutomierz.ui
+package com.fra.plutomierz.ui.components
 
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.MaterialTheme
